+++ .eslintrc
@@ -10,5 +10,6 @@
   "rules": {
   },
   "globals": {
+    "cc",
   },
 }
