+++ index.html
@@ -54,6 +54,13 @@
     <ul id="history-list"></ul>
   </section>
 
+  <script>
+    cc = {};
+  </script>
+  <script src="./src/scripts/historyData.js"></script>
+  <script src="./src/scripts/form.js"></script>
+  <script src="./src/scripts/comment.js"></script>
+  <script src="./src/scripts/branchHistory.js"></script>
   <script src="./app.js?v=4"></script>
 </body>
 </html>
