+++ webpack.config.js
@@ -1,0 +1,9 @@
+// const path = require('path');
+const webpack = require('webpack');
+
+module.exports = {
+  entry: './src/main.js',
+  output: {
+    filename: 'bundle.js',
+  },
+};
